Refactor asynchronous git clone/pull
Based on feedback from @nicd, the way I was starting the task and awaiting was making my code synchronous instead of asynchronous. Starting the all tasks and putting them in a list then awaiting them after they’ve started is the proper way to do asynchronous calls in this repo’s case.